WiP получил данные по загрузке комментария из экстра редюсера
