+++ src/services/api-actions.ts
@@ -8,7 +8,6 @@ import type {OfferCard} from '../types/type-store';
 import {loadOffersNearSlice} from '../store/slices/load-offer-near-slice';
 import type {Comment} from '../types/type-store';
 import {loadCommentsSlice} from '../store/slices/load-comments-slice';
-import {sendCommentsSlice} from '../store/slices/send-comment-slice';
 import {offersFavoriteSlice} from '../store/slices/load-offers-favorite';
 import type {FavoriteStatus} from './type-service';
 import { offersSlice } from '../store/slices/offers-slice';
@@ -63,23 +62,6 @@ const fetchComments = createAsyncThunk<void, string | undefined, Thunk>(
   },
 );
 
-type SendComment = {
-  id: string | undefined;
-  comment: string;
-  rating: number;
-}
-
-
-const sendComment = createAsyncThunk<void, SendComment, Thunk>(
-  'data/sendComment',
-  async ({id, comment, rating}, {dispatch, extra: api}) => {
-
-    const {data} = await api.post<Comment>(`${ApiRoute.Comments}/${id}`, {comment, rating});
-
-    dispatch(sendCommentsSlice.actions.addLoadComment(data));
-  },
-);
-
 const clearErrorAction = createAsyncThunk(
   'clearError',
   () => {
@@ -96,5 +78,4 @@ export {
   fetchOffersFavorite,
   sendFavoriteOffer ,
   clearErrorAction,
-  sendComment
 };
