+++ src/store/slices/send-comment-slice.ts
@@ -1,20 +1,30 @@
 import {createSlice} from '@reduxjs/toolkit';
 import type {Comment,StateComment} from '../../types/type-store';
 import type {PayloadAction} from '@reduxjs/toolkit';
+import { sendComment } from '../../services/thunk/send-comment';
 
 const initialState: StateComment = {
-  comments: null
+  comment: null,
+  error: null,
+  isLoading: null
 };
 
 const sendCommentsSlice = createSlice({
   name: 'loadComment',
   initialState,
-  reducers: {
-    addLoadComment(state, action: PayloadAction<Comment>) {
-
-      state.comments = action.payload;
-    }
-  }
+  reducers: {},
+  extraReducers (builder) {
+    builder
+      .addCase(sendComment.fulfilled, (state, action: PayloadAction<Comment>) => {
+        state.comment = action.payload;
+      })
+      .addCase(sendComment.rejected, (state, action) => {
+        state.error = action.error.message || 'Unknown error';
+      })
+      .addCase(sendComment.pending, (state) => {
+        state.isLoading = true;
+      });
+  },
 });
 
 export {sendCommentsSlice};
